requirements.txt
Django>3
django_rest_knox==4.2.0
djangorestframework==3.14.0
django_filter==23.2
django-structlog==5.1.0
knox==0.1.14

gunicorn==20.1.0

rich==13.3.5 # Used for console logging.

typing_extensions==4.5.0

sentry-sdk[django]==1.23.0
structlog_sentry==2.0.2

psycopg2==2.9.6
pyyaml==6.0
uritemplate==4.1.1
